+++ sorting-algo/ShellSort.cpp
@@ -7,7 +7,6 @@
 
 #include "SortingAlgorithms.hpp"
 
-// Checked
 void ShellSort(int a[], int n)
 {
     for (int interval = n / 2; interval >= 1; interval /= 2)
@@ -26,7 +25,7 @@ void ShellSort(int a[], int n)
     }
 }
 
-void ShellSortCountComp(int a[], int n, int &countCompare)
+void ShellSortCountComp(int a[], int n, ull &countCompare)
 {
     for (int interval = n / 2; ++countCompare && interval >= 1; interval /= 2)
     {
